backend/pyproject.toml
[project]
name = "kg1-backend"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.2",
    "instructor>=1.5.0",
    "langchain>=0.3.1",
    "llama-index-readers-file>=0.2.2",
    "lxml>=5.3.0",
    "openai>=1.50.2",
    "pydantic-settings>=2.5.2",
    "pydantic[email]>=2.9.2",
    "pymupdf>=1.24.10",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.12",
    "requests>=2.32.3",
    "supabase>=2.8.1",
    "torch>=2.4.1",
    "trafilatura>=1.12.2",
    "umap-learn>=0.5.6",
    "uvicorn[standard]>=0.31.0",
]

[tool.uv]
package = false
